Add jpa and h2 and create a PersistenceContext

+++ src/main/java/ch/l0r5/userservice/services/UserService.java
@@ -8,8 +8,8 @@ public interface UserService {
 
     List<User> getAllUsers();
 
-    void addUser(int id);
+    void addUser(long id);
 
-    void removeUser(int id);
+    void removeUser(long id);
 
 }
